fix tag transform, stop atom exhaustion
field checking of incoming tags was converting to atoms to
lookup tag in schema, for permuted tags was leading to atom
exhaustion. Now tag to field check is done by strings.

This includes a breaking change for transform(), the
value must be a binary, and transform default implementation
returns it's argument. A custom transform should always return a
binary.